[FX] Fix submodule naming for subgraph split
[ghstack-poisoned]
